lib/metrics: move metrics-template from metrics-probe module
