+++ lib/metrics/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(METRICS_HEADERS
+    metrics/metrics-template.h
+    metrics/label-template.h
+    PARENT_SCOPE)
+
+set(METRICS_SOURCES
+    metrics/metrics-template.c
+    metrics/label-template.c
+    PARENT_SCOPE)
+
+#add_test_subdirectory(tests)
