+++ lib/metrics/Makefile.am
@@ -1,0 +1,13 @@
+metricsincludedir			= ${pkgincludedir}/metrics
+
+EXTRA_DIST += lib/metrics/CMakeLists.txt
+
+metricsinclude_HEADERS = \
+	lib/metrics/metrics-template.h	\
+	lib/metrics/label-template.h
+
+metrics_sources = \
+	lib/metrics/metrics-template.c	\
+	lib/metrics/label-template.c
+
+#include lib/metrics/tests/Makefile.am
