+++ lib/metrics/label-template.c
