+++ lib/metrics/label-template.h
