+++ lib/metrics/metrics-template.c
